requirements.txt
joblib==0.14.1
matplotlib==3.1.3
nltk==3.4.5
Flask==1.1.2
Pillow==7.1.2
scikit_learn==0.23.1
